Moved eslint shared config into split config packages
no ref
- split node and react eslint config files into separate config packages with different dependencies
- update repo packages to reference shared eslint config packages + remove unnecessary eslint dev deps
- remove async behavior from eslint config since it's no longer needed
