+++ apps/activitypub/package.json
@@ -38,24 +38,15 @@
         "preview": "vite preview"
     },
     "devDependencies": {
+        "@internal/cfg-eslint-react": "workspace:*",
         "@playwright/test": "catalog:",
         "@testing-library/react": "catalog:",
         "@types/jest": "catalog:",
         "@types/react": "catalog:",
         "@types/react-dom": "catalog:",
-        "@eslint/js": "catalog:",
-        "@typescript-eslint/eslint-plugin": "catalog:",
-        "@typescript-eslint/parser": "catalog:",
         "eslint": "catalog:",
-        "eslint-plugin-ghost": "catalog:",
-        "eslint-plugin-react": "catalog:",
-        "eslint-plugin-react-hooks": "catalog:",
-        "eslint-plugin-react-refresh": "catalog:",
-        "eslint-plugin-tailwindcss": "catalog:",
-        "globals": "catalog:",
         "jest": "29.7.0",
         "jsdom": "catalog:",
-        "typescript-eslint": "catalog:",
         "tailwindcss": "catalog:",
         "vite": "catalog:",
         "vitest": "catalog:"
