+++ apps/admin-x-design-system/package.json
@@ -28,6 +28,7 @@
         "@codemirror/lang-html": "6.4.11",
         "@codemirror/state": "6.7.0",
         "@dnd-kit/utilities": "3.2.2",
+        "@internal/cfg-eslint-react": "workspace:*",
         "@storybook/addon-docs": "catalog:",
         "@storybook/addon-links": "catalog:",
         "@storybook/react-vite": "catalog:",
@@ -38,22 +39,13 @@
         "@types/react": "catalog:",
         "@types/react-dom": "catalog:",
         "@types/validator": "catalog:",
-        "@eslint/js": "catalog:",
-        "@typescript-eslint/parser": "catalog:",
         "@vitejs/plugin-react": "catalog:",
         "autoprefixer": "10.5.2",
         "c8": "catalog:",
         "chai": "catalog:",
         "eslint": "catalog:",
-        "eslint-plugin-ghost": "catalog:",
-        "eslint-plugin-react": "catalog:",
-        "eslint-plugin-react-hooks": "catalog:",
-        "eslint-plugin-react-refresh": "catalog:",
-        "eslint-plugin-tailwindcss": "catalog:",
         "glob": "catalog:",
-        "globals": "catalog:",
         "jsdom": "catalog:",
-        "typescript-eslint": "catalog:",
         "lodash-es": "4.18.1",
         "postcss": "catalog:",
         "postcss-import": "16.1.1",
