+++ apps/admin-x-framework/package.json
@@ -83,6 +83,7 @@
         "types"
     ],
     "devDependencies": {
+        "@internal/cfg-eslint-react": "workspace:*",
         "@playwright/test": "catalog:",
         "@testing-library/jest-dom": "catalog:",
         "@testing-library/react": "catalog:",
@@ -90,19 +91,11 @@
         "@types/react": "catalog:",
         "@types/react-dom": "catalog:",
         "@vitejs/plugin-react": "catalog:",
-        "@eslint/js": "catalog:",
         "@vitest/coverage-v8": "catalog:",
         "c8": "catalog:",
         "eslint": "catalog:",
-        "eslint-plugin-ghost": "catalog:",
-        "eslint-plugin-react": "catalog:",
-        "eslint-plugin-react-hooks": "catalog:",
-        "eslint-plugin-react-refresh": "catalog:",
-        "eslint-plugin-tailwindcss": "catalog:",
         "glob": "catalog:",
-        "globals": "catalog:",
         "jsdom": "catalog:",
-        "typescript-eslint": "catalog:",
         "msw": "catalog:",
         "type-fest": "catalog:",
         "typescript": "catalog:",
