+++ apps/admin-x-settings/package.json
@@ -73,6 +73,7 @@
         "validator": "catalog:"
     },
     "devDependencies": {
+        "@internal/cfg-eslint-react": "workspace:*",
         "@playwright/test": "catalog:",
         "@testing-library/jest-dom": "catalog:",
         "@testing-library/react": "catalog:",
@@ -84,19 +85,9 @@
         "@types/react": "catalog:",
         "@types/react-dom": "catalog:",
         "@types/validator": "catalog:",
-        "@eslint/js": "catalog:",
-        "@typescript-eslint/eslint-plugin": "catalog:",
-        "@typescript-eslint/parser": "catalog:",
         "@vitest/coverage-v8": "catalog:",
         "eslint": "catalog:",
-        "eslint-plugin-ghost": "catalog:",
-        "eslint-plugin-react": "catalog:",
-        "eslint-plugin-react-hooks": "catalog:",
-        "eslint-plugin-react-refresh": "catalog:",
-        "eslint-plugin-tailwindcss": "catalog:",
-        "globals": "catalog:",
         "jsdom": "catalog:",
-        "typescript-eslint": "catalog:",
         "tailwindcss": "catalog:",
         "vite": "catalog:",
         "vitest": "catalog:"
