+++ apps/admin/eslint.config.js
@@ -8,7 +8,7 @@ import { globalIgnores } from 'eslint/config'
 import noRelativeImportPaths from 'eslint-plugin-no-relative-import-paths'
 import ghostPlugin from 'eslint-plugin-ghost';
 
-import {correctnessRules, shadeLayeredImportsRule, strictLinterOptions} from '../../eslint.shared.mjs';
+import {correctnessRules, shadeLayeredImportsRule, strictLinterOptions} from '@internal/cfg-eslint';
 
 const noHardcodedGhostPaths = {
   meta: {
