+++ apps/admin/package.json
@@ -42,6 +42,7 @@
     },
     "devDependencies": {
         "@eslint/js": "catalog:",
+        "@internal/cfg-eslint": "workspace:*",
         "@tailwindcss/vite": "catalog:",
         "@tanstack/react-query": "catalog:",
         "@testing-library/jest-dom": "catalog:",
@@ -56,6 +57,7 @@
         "@vitejs/plugin-react": "catalog:",
         "@vitest/browser-playwright": "catalog:",
         "eslint": "catalog:",
+        "eslint-plugin-ghost": "catalog:",
         "eslint-plugin-no-relative-import-paths": "1.6.1",
         "eslint-plugin-react-hooks": "catalog:",
         "eslint-plugin-react-refresh": "catalog:",
