+++ apps/portal/package.json
@@ -42,7 +42,7 @@
     },
     "devDependencies": {
         "@doist/react-interpolate": "2.2.2",
-        "@eslint/js": "catalog:",
+        "@internal/cfg-eslint-react": "workspace:*",
         "@sentry/react": "catalog:",
         "@testing-library/jest-dom": "catalog:",
         "@testing-library/react": "catalog:react17",
@@ -54,14 +54,9 @@
         "cross-fetch": "catalog:",
         "dompurify": "catalog:",
         "eslint": "catalog:",
-        "eslint-plugin-ghost": "catalog:",
-        "eslint-plugin-i18next": "6.1.5",
-        "eslint-plugin-react": "catalog:",
-        "globals": "catalog:",
         "jsdom": "catalog:",
         "react": "catalog:react17",
         "react-dom": "catalog:react17",
-        "typescript-eslint": "catalog:",
         "vite": "catalog:",
         "vite-plugin-svgr": "catalog:",
         "vitest": "catalog:"
