+++ apps/shade/eslint.config.js
@@ -1,7 +1,7 @@
-import {reactAppConfig} from '../../eslint.shared.mjs';
+import {reactAppConfig} from '@internal/cfg-eslint-react';
 
 // Shade lints `scripts/` alongside `src/` (one-off build/codegen scripts).
-export default await reactAppConfig({
+export default reactAppConfig({
     tailwindCssPath: `${import.meta.dirname}/../admin/src/index.css`,
     ignores: ['dist/**/*', 'storybook-static/**/*'],
     srcGlobs: ['src/**/*.{js,ts,cjs,tsx}', 'scripts/**/*.{js,ts,cjs,tsx}'],
