+++ apps/shade/package.json
@@ -82,29 +82,19 @@
         "preflight.css"
     ],
     "devDependencies": {
+        "@internal/cfg-eslint-react": "workspace:*",
         "@storybook/addon-docs": "catalog:",
         "@storybook/addon-links": "catalog:",
         "@storybook/react-vite": "catalog:",
         "@tailwindcss/postcss": "catalog:",
         "@testing-library/react": "catalog:",
         "@types/node": "catalog:",
-        "@typescript-eslint/eslint-plugin": "catalog:",
-        "@typescript-eslint/parser": "catalog:",
         "@vitejs/plugin-react": "catalog:",
         "@vitest/coverage-v8": "catalog:",
-        "@eslint/js": "catalog:",
         "c8": "catalog:",
         "eslint": "catalog:",
-        "eslint-plugin-ghost": "catalog:",
-        "eslint-plugin-react": "catalog:",
-        "eslint-plugin-react-hooks": "catalog:",
-        "eslint-plugin-react-refresh": "catalog:",
-        "eslint-plugin-storybook": "10.4.6",
-        "eslint-plugin-tailwindcss": "catalog:",
         "glob": "catalog:",
-        "globals": "catalog:",
         "jsdom": "catalog:",
-        "typescript-eslint": "catalog:",
         "postcss": "catalog:",
         "remark-gfm": "4.0.1",
         "storybook": "catalog:",
