+++ apps/sodo-search/eslint.config.js
@@ -1,9 +1,9 @@
-import {reactAppConfig} from '../../eslint.shared.mjs';
+import {reactAppConfig} from '@internal/cfg-eslint-react';
 
 // LEGACY: vanilla JS (no TypeScript). sodo-search should migrate to TS — small
 // surface (~6 components) so the migration is 1–2 days. Until then, the
 // typescript: false flag is the escape hatch.
-export default await reactAppConfig({
+export default reactAppConfig({
     typescript: false,
     reactRefresh: false,  // bundled as UMD for theme distribution
     sortImports: true,
