+++ ghost/admin/package.json
@@ -42,6 +42,7 @@
     "@faker-js/faker": "10.4.0",
     "@glimmer/component": "1.1.2",
     "@html-next/vertical-collection": "3.0.0",
+    "@internal/cfg-eslint": "workspace:*",
     "@sentry/ember": "7.120.3",
     "@sentry/integrations": "7.114.0",
     "@sentry/replay": "7.116.0",
