+++ ghost/i18n/package.json
@@ -41,12 +41,10 @@
     "locales"
   ],
   "devDependencies": {
-    "@eslint/js": "catalog:",
+    "@internal/cfg-eslint": "workspace:*",
     "@vitest/coverage-v8": "catalog:",
     "eslint": "catalog:",
-    "eslint-plugin-ghost": "3.5.0",
     "fs-extra": "catalog:",
-    "globals": "17.7.0",
     "glob": "catalog:",
     "i18next-parser": "9.4.0",
     "vitest": "catalog:"
