+++ koenig/kg-card-factory/package.json
@@ -41,6 +41,7 @@
     "@eslint/js": "catalog:",
     "@types/node": "catalog:",
     "@vitest/coverage-v8": "catalog:",
+    "eslint-plugin-ghost": "catalog:",
     "simple-dom": "1.4.0",
     "typescript": "catalog:",
     "typescript-eslint": "catalog:",
