+++ koenig/kg-default-nodes/package.json
@@ -57,6 +57,7 @@
     "@types/luxon": "3.7.2",
     "@types/node": "catalog:",
     "@vitest/coverage-v8": "catalog:",
+    "eslint-plugin-ghost": "catalog:",
     "html-minifier": "4.0.0",
     "prettier": "3.9.4",
     "typescript": "catalog:",
