+++ koenig/kg-html-to-lexical/package.json
@@ -39,6 +39,7 @@
     "@eslint/js": "catalog:",
     "@types/jsdom": "catalog:",
     "@vitest/coverage-v8": "catalog:",
+    "eslint-plugin-ghost": "catalog:",
     "typescript": "catalog:",
     "typescript-eslint": "catalog:",
     "vitest": "catalog:"
