+++ koenig/kg-lexical-html-renderer/package.json
@@ -41,6 +41,7 @@
     "@eslint/js": "catalog:",
     "@types/jsdom": "catalog:",
     "@vitest/coverage-v8": "catalog:",
+    "eslint-plugin-ghost": "catalog:",
     "jsdom": "catalog:",
     "prettier": "3.9.4",
     "typescript": "catalog:",
