+++ koenig/kg-markdown-html-renderer/package.json
@@ -43,6 +43,7 @@
     "@types/node": "catalog:",
     "@types/semver": "7.7.1",
     "@vitest/coverage-v8": "catalog:",
+    "eslint-plugin-ghost": "catalog:",
     "typescript": "catalog:",
     "typescript-eslint": "catalog:",
     "vitest": "catalog:"
