+++ koenig/kg-utils/package.json
@@ -39,6 +39,7 @@
     "@types/node": "catalog:",
     "@types/semver": "7.7.1",
     "@vitest/coverage-v8": "catalog:",
+    "eslint-plugin-ghost": "catalog:",
     "typescript": "catalog:",
     "typescript-eslint": "catalog:",
     "vitest": "catalog:"
