+++ packages/_template/eslint.config.mjs
@@ -1,3 +1,3 @@
 import {nodeLibConfig} from '@internal/cfg-eslint';
 
-export default await nodeLibConfig();
+export default nodeLibConfig();
